11: finalize inital setup for Vehicle LiisrView
