+++ android/VehicleDiary/app/src/main/java/com/cool/element/vehiclediary/MainActivity.kt
@@ -7,12 +7,11 @@ import androidx.activity.enableEdgeToEdge
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import com.cool.element.vehiclediary.domain.Vehicle
-import com.cool.element.vehiclediary.presentation.VehicleListView
+import com.cool.element.vehiclediary.presentation.vehicle.VehicleListView
 import com.cool.element.vehiclediary.ui.theme.VehicleDiaryTheme
 
 class MainActivity : ComponentActivity() {
