+++ android/VehicleDiary/app/src/main/java/com/cool/element/vehiclediary/domain/Vehicle.kt
@@ -41,7 +41,7 @@ data class Vehicle (
             vin = "Test vin",
             licensePlate = "Test license plate",
             purchaseDate = java.time.LocalDateTime.now(),
-            image = "Test image",
+            image = null,
             odometer = 1000,
             createdAt = java.time.LocalDateTime.now(),
             updatedAt = java.time.LocalDateTime.now(),
@@ -60,7 +60,7 @@ data class Vehicle (
                         vin = "Vehicle $i Vin",
                         licensePlate = "Vehicle $i License Plate",
                         purchaseDate = java.time.LocalDateTime.now(),
-                        image = "Vehicle $i Image",
+                        image = null,
                         odometer = 1000,
                         createdAt = java.time.LocalDateTime.now(),
                         updatedAt = java.time.LocalDateTime.now(),
