+++ android/VehicleDiary/app/src/main/java/com/cool/element/vehiclediary/presentation/vehicle/VehicleListView.kt
@@ -1,0 +1,50 @@
+package com.cool.element.vehiclediary.presentation.vehicle
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.items
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.material3.*
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import coil.compose.rememberAsyncImagePainter
+import com.cool.element.vehiclediary.domain.Vehicle
+
+@Composable
+fun VehicleListView(
+    vehicles: List<Vehicle>,
+    modifier: Modifier = Modifier
+) {
+    LazyColumn(
+        modifier = modifier
+            .fillMaxSize()
+            .padding(16.dp)
+    ) {
+        items(vehicles) { vehicle ->
+            VehicleRow(vehicle = vehicle)
+            HorizontalDivider(
+                modifier = Modifier.padding(vertical = 8.dp),
+                thickness = 1.dp,
+                color = Color.Gray
+            )
+        }
+    }
+}
+
+@Preview(showBackground = true)
+@Composable
+fun PreviewVehicleListView() {
+    val vehicles = Vehicle.sampleList
+
+    VehicleListView(vehicles = vehicles)
+}
